Setup separate application.property profiles
When running a workflow that needs access to the database (such as testing), you to need to tell your workflow that you want to run with the testing profile. By default, its the dev profile which doesn't have a username and password in it.
For example, if you're using a jar file for testing, you would add the command: "java -jar -Dspring.profiles.active=dev {cosc310project}.jar" in the workflow